Added config params for ifdef consts, moved device sel. class decl. to new file
Added config parameters for __SYCL_COMPILER_VERSION cutoffs used throughout.
Moved declarations of dpctl_device_selector and other derived classes to a
dedicated .hpp file.
